sequences: move SequenceForm to Views/
Per the views/components rule (#4324): single-use components belong
in app/views/controllers/<controller>/.

  Components::SequenceForm
    -> Views::Controllers::Sequences::Form
  app/components/sequence_form.rb
    -> app/views/controllers/sequences/form.rb
  test/components/sequence_form_test.rb
    -> test/views/controllers/sequences/form_test.rb

Two callers: edit.html.erb and new.html.erb render the form
directly. ModalTurboForm also dispatches to this form dynamically
for sequence-create modals on observation pages — picks up the new
path automatically via the dual-lookup from #4328.

Mechanical refactor. No behavior changes.
